get session log working again
install the tool in the dockerfile, as user thopter, using uv env vars that place everything in /opt so that the homedir mount doesn't conflict with uv's tool installation. this means creating the thopter user in the dockerfile instead of in the init script initially.

also work around the cache bug in the project by specifically clearing out html comments it creates that prevent recreation of the html files. before i was deleting the html files entirely, which created a period where they wouldn't render and you'd get the plain python webserver directory listing